feat: Update mapdata and markers for DeeDeeRescue
Update the mapdata.lua file to change the NetHash and ObjectsHash values for the uhSeTtP map. Also, update the markers.debug.lua file to change the group name and marker name from "DeeDeeMissing" to "DeeDeeRescue". This change is necessary to reflect the updated functionality of the mod.

fix: Adjust map orientation and height map values

Adjust the map orientation and height map values to ensure accurate representation of the terrain in the uhSeTtP map.

feat: Update mod version to 1.1.413

Bump up the mod version to 1.1.413 to reflect the latest changes made in the mod. This update includes a fix for B1 being accessible from the east. Additionally, external links have been added to the mod metadata, pointing to the GitHub repository for the mod.

init

Initial commit